panopticon/pyproject.toml
[tool.poetry]
name = "panopticon"
version = "0.1.0"
description = "Weather data dashboard"
authors = ["Thomas Steffes <tom@soteris.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.8.1"
pandas = "^1.5.3"
dash-bootstrap-templates = "^1.0.8"
click = "^8.1.3"
sqlalchemy = "^2.0.4"
psycopg2 = "^2.9.5"
paho-mqtt = "^1.6.1"
dash-daq = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
